Don't crash on loading no-longer supported savestates
This was an oversight in the serialization framework:
* We did already check for 'unknown polymorphic types' in
  PolymorphicInitializerRegistery::init().
* But we did not in PolymorphicLoaderRegistery::load().
Both should be symmetric. This patch adds the check in the latter.

The type (=string with the name of the type) can be missing when:
* A savestate was manually edited.
* A newer openMSX version removes some old functionality.

And in the next patch removing functionality is exactly what we'll do.
See there for more details.