ImGuiMessages: possibility to show messages on screen with auto fade-out
This is a 3rd possibility to show messages (configurable via "Settings >
Misc > Configure messages ..."). And it's the default for info messages.

These messages are shown in the top left corner of the main openMSX
window. And they fade out after a configurable amount of time (default 5
seconds).

Because, compared to the old OSD messages, these are shown in a smaller
font (actually the default ImGui font size, having lots of diffent font
sizes is not efficient with Dear ImGui), these messages first flash
bright to draw attention.

When upgrading to this version you may need to configure your message
settings before this is used. Or delete your imgui.ini file, then it's
the default for info messages.