Fix clang warning: -Wunused-function
The functions memset_16() and memset_16_2() are no longer needed after
we removed the software video rendering path.

These functions were still called in the code, but only from inside a
dead "if-constexpr" branch. Apparently gcc doesn't give warnings about
this. You can make a point for both the gcc and the clang behavior.