Fix clang warning: -Wimplicit-int-conversion
Most of these warnings are harmless, and are only there because of c/c++
integer promotion rules, but the value can not actually get out of
range.

In a few places I added extra input-string validation (for stuff like
"button1000" but we 'only' support 256 joystick buttons).

After this the clang build is warning free again. Except for some
warnings in the the src/3rdpart subdirectory. Ideally we only compile
3rdparty files with standard warning settings, or without any warnings.
But unfortunately that's not supported in our homegrown build system.