RS232Net: eliminate 'socket_address' member variable
This variable was only used to temporarily store the parsed form of
'rs232NetAddressSetting'. And that is only needed inside the
'plugHelper()' method. So instead of being a member variable, it's
better to make it a local variable of 'plugHelper()'.

Also renamed 'network_address_generate()' to 'parseNetworkAddress()' and
made it a free function (instead of a member function). Because after
some slight changes it doesn't need any info from the RS232Net class
anymore.

The motivation for eliminating member variables is to reduce the
possible 'states' an object can be in. And then in particular eliminate
possibly inconsistent states. In this case 'socket_address' was only
in-sync with 'rs232NetAddressSetting' at a specific moment in time. And
then keeping it around for longer than that time can only cause
confusion.

I removed the comment about the union-stuff (about members of a union
not starting at the same address) for these reasons:
* Comments should explain what, in particular why, the code is doing
  something. It should not explain a language feature.
* But the comment is also plain wrong, or at least outdated. See here
  for more info:
    https://stackoverflow.com/questions/43747487/does-the-c17-standard-guarantee-that-the-address-of-a-union-is-the-same-as-the