RS232Net: document and fix some threading/locking issues
Document which member variables are read/written by which threads (the
main emulation thread and the helper thread that reads data from the
network).

Shared variable (between the above two reads) where at least one of the
threads write to the variable need:
* either hold a lock while doing either read or write (we use the
  'mutex' variable for this).
* The variable must be a std::atomic<>. Note: I've made 'SOCKET' also a
  std::atomic. That's fine on Linux (there SOCKET is an 'int'). I think
  it should also be fine on Windows, but we'll see what the automatic
  build jobs do with it.

In a few places we did too many checks for 'sockfd ==
OPENMSX_INVALID_SOCKET'. I've now made this a pre-condition of the
'net_putc()' and 'net_getc()' functions. So these functions themselves
no longer check this condition (to avoid duplicating this check).
Instead the caller must already have checked this.