ImGuiUtils: accept new setting value after widget looses focus
Requested by durandalwoz:
    Is it possible for text fields to accept the text already typed in
    when clicked out of focus? Currently the new text is only accepted
    after pressing enter. Otherwise it reverts to the previous text

For string settings we don't want to change the setting value after
every key-press. Therefor we only accepted the new value after the user
pressed enter. Though when you click away (close the window, or click in
some other setting to edit) it would still be nice to accept the new
value at that point. (And aborting the edit is still possible via the
'ESC' key).

It seems that Dear ImGui has a function especially for this purpose:
    ImGui::IsItemDeactivatedAfterEdit()
Easy fix with that function.