ImGuiTrainer: when opening ComboBox, scroll to selected item
The difficulty was that (for efficiency) in a very long list we use
ImGuiListClipper to not draw all (invisible) items. But that also means
we skip the logic that would otherwise handle the
ImGui::SetItemDefaultFocus() stuff.

Fixed by forcing the selected item to be processed (even if it is
invisible). This advantage of this approach is that we now, on each
frame, still do a string-comparison against the full list (I didn't find
an easy way to avoid this). Before we only needed to compare the visible
items.

See also this discussion:
    https://github.com/ocornut/imgui/issues/5995