RS232Net: fix possibles crash (on reverse)
In commit 4eee1c72e3ab1 I added a precondition for net_putc():
   sockfd != OPENMSX_INVALID_SOCKET
The motivation was to avoid double checking this condition. Because
there was code that checked this before calling net_putc(), and then
net_putc() repeated that same check internally.

This caused a bug in code like this:
    if (sockfd != OPENMSX_INVALID_SOCKET) {
        net_putc(IP232_MAGIC);
        net_putc(IP232_DTR_LO);
    }
At first sight this looks fine. However in case there's an error inside
net_putc(), it will close the connection and set sockfd=OPENMSX_INVALID_SOCKET.
So the first call to net_putc() is fine, but the second one can fail
(trigger an assertion).

Instead of adding a check before the 2nd net_putc() call, I changed this
function from:
    bool net_putc(char b);
to
    bool net_put(std::span<const char> buf);
So instead of sending one byte at a time, the function can now send a
whole buffer. And then in the above sequence we would send two bytes at
once. (Sending a single byte also remains possible of course).

This fixes the crash, and as a bonus it's also a bit more efficient.

--

There was a second problem. On reverse we don't to restore the state of
the RS232Net pluggable. Trying to do this is pointless because we anyway
can't restore the state of the server we're connected to.

Though because of this (after reverese) we can get in the situation that
the RS232Net is plugged in, but we did not start the receiver-thread.
And then later on in RS232Net::unplugHelper() we call 'thread.join()' on
a thread that hasn't been started yet. But that throws an exception.

Easy to fix once the problem was identified.