RS232Net: 'properly' use Poller
This patch fixes a few problems:

1) Poller objects could only be used once: once aborted they stayed in
the aborted-state. This affected RS232Net, but also RS232Tester and
MidiInReader. (It did not affect CliConnection, for which Poller was
originally created).

This patch adds Poller::reset() to restore the object to the non-aborted
state.

2) The RS232Net receiver thread did select() with a zero-timeout. So
basically just a busy-loop using polling. And indeed as soon as the
RS232Net device was plugged in, there was an extra thread using 100%
CPU. After this patch, CPU usage does not noticably increase. See next
point for how this was fixed.

3) The Poller object wasn't properly used. Before this patch we
basically only used 'abort()' and 'aborted()' (but if that's the only
thing we need, then std::atomic<bool> would be a better solution). The
main purpose of Poller is the wait until a file/socket becomes readable
without using 100% CPU and do this in a portable(*1) and abort-able way.

So calling 'Poller::poll(int fd)' blocks until 'fd' becomes readable OR
until some other thread calls 'Poller::abort()' on the same object.

(*1) It's portable between unix variant, but for Windows we instead rely
on the fact that closing a socket an which another thread is doing a
blocking read, will unblock that read operation.