Improvements to RS-232 port 0x82 read values
Unused bits for port 0x82 are left floating, therefore reading as 1 for the processor.
MSXRS232::readstatus() modified accordingly: now we start with 0xFF and mask out any active bit.

A RS232 port with no plugged in device will return a different value for it's input control lines depending on the hardware implementation:
Floating or pulled down lines will be read as 1 for the processor. (Most common implementation)
Pulled up lines will read as 0 for the processor. (So far only the SVI-738 is know to have this implementation)

Input control lines methods in RS232Device class modified to return `false` by default (floating lines implementation).

Added getDSR() and getCTS() methods to RS232Net and RS232Tester to make sure the lines are set to their correct state when a device is plugged in.

Added rs232_pullup tag (set to true) in the config files for the SVI-738 and it's variants.

TODO:
Pass a DeviceConfig object to RS232Device to read the value of the rs232_pullup setting.
Negate that value and use it as return value for the control lines method (instead of the fixed `false`)