Simulate unplugged RS-232 input control lines with configurable 'pull-up'/'pull-down' behaviour
RS232Device.cc input control lines methods now return a boolean optional.
An empty return value means the input is unplugged, and it should default to the value of the configuration tag 'rs232_pullup'.
This configuration tag defaults to false and reflects the hardware implementation of the RS-232 port.
As of now only the SVI-738 is known to need this tag set to true.