gl::vec: allow to access components via 'v.x'
'gl::vecN' is an N-component mathematical vector (N=2,3,4).
Accessing the individual components can be done via operator[]. For
example:
    vec4 v(1, 2, 3, 4);
    v[0] = 10;
    f(v[2]);

Dear ImGui has similar vector types called ImVec2 and ImVec4. (Both our
and ImGui types have different use cases, so we need both). Accessing
components in a ImVec4 is done via .x, .y, .z and .w members. For
example:
    ImVec4 v(1, 2, 3, 4);
    v.x = 10;
    f(v.z);

In some of our imgui/ code we need a mix of both gl::vec2 and ImVec2 (or
gl::vec4). Including accessing components. And then it's annoying that
both use a different syntax. The operator[] syntax is more useful for
generic vector code (e.g. when looping over the components). The .x
syntax is technically more limited, but it can be more descriptive.

This patch extends our gl::vec type so that both syntaxes are possible.
Internally this is implemented via a union. Where appropriate this patch
also switches to the new syntax.