ImGuiConsole: improve drawing of console input line
Because Dear ImGui does not (currently) support multi-color text input
(we need this for syntax highlighting in the console), we use the
following hack:
* We first use a general ImGui::InputText() widget. This takes care of
  all input handling and updates the (in-progress) command string.
* Then we overdraw the text with the correct colors.

This used to work fine. At least when we were using the default builtin
font (ProggyClean.ttf). This is a pixel-perfect bitmap font, with only
fully opaque pixels. This mean if we overdraw this font, all previously
drawn pixels are fully replaced with a different color.

Later we switched the default console font to 'DejaVuSansMono.ttf'. Here
pixels do have an alpha-value. And this means if you overdraw the font,
part of the first draw-pass remains visible (the characters look weird:
too bright and too bold).

In this patch we fix this by drawing the first pass fully transparent.
Unfortunately this includes the text-cursor, so that one disappears now.
And to fix that we also need to (over)draw the text cursor ourselves.
Doing that (also) requires peeking in some Dear ImGui internal APIs /
data-structures (but we already had to do that for overdrawing the
colored text itself, see comment in the code for more details).

After this the input command line looks much cleaner again.