gl::vecN: fix(?) windows after commit 1dcf317029838
Commit 1dcf317029838 extended the gl::vecN interface so that stuff like
  vec2 pos = ...
  myFunction(pos[0], pos[1]);
could also be written like:
  myFunction(pos.x, pos.y);
This often leads to more descriptive code (and a bit more compact).

The implementation of 1dcf317029838 relied on unions. Internally the
VecN type stores:
* an array of N elements
* overlayed with N named scalars 'x', 'y', ...

This worked fine on gcc/clang, but not on windows (msvc). I now
*believe* this is because this trick relied on a gcc/clang extension,
but this isn't correct according to the C++ standard.

I believe gcc/clang promise that when two members of a union (possibly
those members are themselves structs or arrays) have the same type on
the same offset, then that particular type can be accessed via either
member. For example:
    union {
        int a[2];
        struct { int x, y; }
    };
    // 'a[0]' is interchangable with 'x', same for 'a[1]' and 'y'

Though that's stricter than what the c++ standard requires. It states:
   It is undefined behavior to read from the member of the union that
   wasn't most recently written.
In other words, in the example above, when 'x' was last written, it's
only allowed to use this value again via 'x', not via 'a[0]'.

--

I fixed(?) this by removing the complicated union/struct stuff. Instead
the VecN class(es) now have plain 'x', 'y', 'z', ... data members. This
has two disadvantages:
* operator[] now can no longer index into an array. Instead there's an
  if-then-else-chain:
    if (i == 0) return x;
    if (i == 1) return y;
    ...
  Usually operator[] is called with a constant index, or with a
  loop-index from within a loop which usually gets unrolled. And then
  the inliner/optimizer can still generate good code.
  Though possibly in some cases we now have worse code. This isn't too
  bad because these operations typically aren't used in performance
  critical code.
* It's no longer possible to write 1 generic 'vecN<N, T>' template which
  handles all sizes. Instead we now have 3 specialization of the base
  template for N=2, N=3 and N=4. This is a bit more code with some
  duplication.

If needed (in the future) we could mitigate these problems:
* We could manually specialize / unroll the vector operations (those
  that now use operator[]). This reduces the reliance on the
  inliner/optimizer.
* We could use CRTP to reduce some of the code duplication. Though the
  current code is not too bad, and introducing CRTP might not make the
  code simpler.