DebuggableEditor: bugfix update changing field when not editing
This recently broke in commit 084d90b857729. Thanks to bas for reporting
(and discovering this so quickly).

In the memory view, when the cursor is an a changing field (e.g. 0xfc9e
JIFFY), and when the edit field is deactivated (e.g. click somewhere
else), then the (changing) value is no longer updated.

Now I understand why the original code went back to the 'NONE' state
(see previous commit message) ;-)