Disassembly window: bug fix when breakpoint window wasn't open
This is a follow-up on c76da528b9a. Thanks to 'bas' for reporting this
bug.

Clicking in the 'bp' column in the disassembly view would create a new
breakpoint, but it was only shown once you opened the "Breakpoint
window". (In my tests yesterday I always had this window open).

The problem is duplication of state:
* Originally only the openMSX CPU emulation code stored information
  about breakpoints. This is a compact representation, it contains only
  enabled breakpoints and it's always sorted on address.
* When we added the GUI breakpoint window we needed more info: we need
  to retain info about disabled breakpoints, we want to be able to sort
  on other fields, we want to keep the address-string instead of only
  the parsed numeric value (e.g. CHMOD vs 0x005f).

So currently the information about breakpoints is duplicated. And the
bug was that the Breakpoint window DOES synchronize these two, but the
disassembly window did NOT.

Maybe in the future we could reduce this duplication? Although I'm not
sure yet how to do that (preferably without introducing GUI specific
knowledge in the CPU emulation code).