ImGuiSpriteViewer: fixed off-by-one error
To simplify the sprite-mode-2 rendering, the list of visible sprites
ends with a sentinel value. This sentinel is implemented as a sprite
with attrib=0, the other values (x and pattern) are left uninitialized.

Though we had an off-by-one bug. When (backwards) iterating over the
visible sprites, we wrongly included this sentinel.