Forwarding constructor shouldn't hide copy-constructor
In c++ a forwarding constructor is one that "forwards" it's constructor
parameters, typically to construct some member variable:
   template<typename T>
   MyClass(T&& T) : foo(std::forward<T>(t)) {}
in other words, this constructor accepts any type of parameter and
simply passes it to the constructor of the 'foo' member (e.g. foo might
be a std::string and a std::string can be constructor from various
types).

The copy constructor looks like this:
    MyClass(const MyClass&) { ...
So it takes an object of the same type. And because it won't change the
original object it's accepted as a *const* reference.

Now when you make a copy like this:
    MyClass c1;     // note: a non-const object
    MyClass c2(c1);
we actually pass a non-const object. That's fine: the copy constructor
should accept both const and non-const objects. HOWEVER the above
forwarding constructor can be instantiated with T=MyClass to form
    MyClass(MyClass&)
and that's actually a better match than the copy constructor. But that
was not the intention.

"sonarcloud" flagged 6 constructors in our code where we did something
like this. And for one of those we actually already hit this problem and
added a workaround (workaround was to add an extra non-canonical
copy-constructor that takes a non-const-reference, but that's a hack).

The better solution is to use the (C++20) "requires" annotation to
prevent the forwarding constructor from being instantiated for types
that are the same as the class type (see patch, that should make it
clear).