HotKey: allow to give hotkeys lower priority than ImGui
Heavily based on:
  https://github.com/openMSX/openMSX/pull/1593
Thanks a lot to pvmm for the initial code ideas!! And especially also
for starting this work, and doing several iterations of updates. Without
his effort this work would like still be stalled for many weeks/months.

See the same PR for a lot of discussion.

In short:
* The bind command now takes an optional "-msx" flag. Such binding are
  only active when the main msx window has focus (so when no ImGui
  window has focus).
* Currently this is (only) used for the PgUp/PgDown bindings (reverse).
  So now PgUp/PgDown becomes usable in ImGui.