Remove dead code after 'UNREACHABLE'
In openMSX we've long had the 'UNREACHABLE' macro. This is used to
indicate that a certain code path can never be reached. For example:
    switch (x) {
        case 2: ...
        case 4: ...
        case 8: ...
        default: UNREACHABLE;
    }
When build with asserts enabled, UNREACHABLE expands to 'assert(false)'.
Without asserts, depending on the compiler (gcc/clang/msvc) it expands
to a compiler specific builtin-function indicating that this code path
cannot be reached, and then the compiler will indeed optimize the code
based on that assumption. (C++23 added std::unreachable(), in the future
we can replace our macro with that).

Nowadays compiler understand that the code-flow stops at UNREACHABLE, in
the past this wasn't always the case. So we still had code like this
(note the dummy return statement):
    int someFunction() {
        switch (x) {
            case 2: return 100;
            case 4: return 200;
            case 8: return 300;
            default: UNREACHABLE; return 0; // dummy value to suppress warning
        }
    }

This patch removes all such dummy code. I verified that indeed all
compilers, with or without asserts enabled, still compile without
warnings on this simplified construct.

This cleanup was triggered after a report from 'sonarcloud'. (Though it
only complained about 11 cases, but this patch cleans up many more
cases).