Fix undefined behavior related to unions
Reported by 'sonarcloud'. Strictly speaking the report is correct, it is
undefined behavior. Though in practice all compilers generated correct
code anyway (and I _think_ gcc even guarantees this correct behavior, as
an extension on the c++ standard).

In more detail. We had code like this:
    SDL_Event evt = {};
    e.key.type = SDL_KEYUP;
    e.key.timestamp = SDL_GetTicks();
    e.key.state = SDL_RELEASED;
The intention was that the first statement would initialize (=set to
zero) the full enum. And then the later statements initialize some of
the 'key' members of the union.

This is wrong for two reasons:
* The statement 'SDL_Event evt = {}' initializes the *first* member of
  the union (not the full union). The first member happens to be 'type'
  (and this is the common head with all other union members), so
  actually only the 'e.key.type' member was set to zero.
* According to the c++ standard it's only allowed to access the 'active
  member' of a union. So something like this is not allowed:
    union U { int i; float f; };
    U u; u.i = 123; use(u.f);
  In the example above we have something similar: we first initialize
  the first member of the union (evt.type). But then we access the
  'evt.key' member. It's not clear to me whether writing evt.key.type
  is undefined behavior or that it actives the 'key' member. In any case
  'sonarcloud' complained about it, so I assume it's not valid.

The fix was fairly easy. The above example becomes:
    SDL_Event evt;
    e.key = SDL_KeyboardEvent{};
    e.key.type = SDL_KEYUP;
    e.key.timestamp = SDL_GetTicks();
    e.key.state = SDL_RELEASED;
Note that the first member does not yet initialize the union. So it does
not yet select an active member. In the second statement we initialize
the 'key' member (so now this gets allo fully initialized). And all
later statements also only access the 'key' member. So there's no mixing
of (active) members anymore.