Replace (most) reinterpret_cast with std::bit_cast
std::bit_cast is a new c++20 feature (and only available starting from
gcc-11).
    https://en.cppreference.com/w/cpp/numeric/bit_cast

It's a safer version of reinterpret_cast. And safer because it allows
less than what reinterpret_cast allows.

In particular std::bit_cast is only allowed when:
* Casting between two types that have the same size.
* Both types must be trivially copyable.

For example you can bit_cast between different pointer types (at least
on most platforms all pointers have the same representation). But you
cannot bit_cast between a pointer and an 'int' (on a typical 64 bit
platform pointers are 64-bit but 'int' is 32 bit). Instead a pointer
must be bit_cast to either 'intptr_t' or 'uintptr_t' (in a few places we
made this mistake ptr->int, although it seemed to work fine in
practice).

So, if possible, reinterpret_cast should be replaced with bit_cast. But
of course it's still better to avoid the cast completely, if possible.