Replace redundant type with 'auto'
E.g. when the same type is specified twice in the same line:
    VDP* vdp = dynamic_cast<VDP*>(device);
->
    auto* vdp = dynamic_cast<VDP*>(device);

Our code already was quite good in following this rule. But for some
reason I didn't do this for types that were only 3 or 4 chars long (so
stuff like 'int', 'char', 'VDP', ...). Yes, for those types, replacing
the duplicate type with 'auto' makes the line 1 character longer. But
that's not a good reason to introduce an inconsistency.

Change triggered by 'sonarcloud' warnings.