Fix crash on loading layout, after 93e82a566b5
In commit 93e82a566b5 we postponed removing elements from the 'parts'
vector, to avoid iterating over a changing vector.

Though now the problem is that (in one specific loop) we now visit
objects that are already deleted.

Now I changed the code to:
* In unregisterPart() replace the to be remove part with nullptr. This
  avoids the problem of iterating over a changing vector (the content of
  the vector changes (that's fine), but the vector does not grow or
  shrink).
* In some loops we now check that an element has been replaced by
  nullptr before using it.
* In updateParts() we remove all nullptrs from the vector.

This fixes the bug, and it's also a bit more efficient.