Set block size to 0 for non-mapper ROMs
Triggered by this PR:
  https://github.com/openMSX/openMSX/pull/1600

Thanks to pvmm for pointing out the problem and proposing a solution.
Thanks to MBilderbeek for suggesting an alternative solution.
I merely verified the idea and pushed it.

Background info:
Our `RomPlain` and `RomPageNN` classes currently inherit from
`Rom8kBBlocks`. That should be only an implementation detail (it allowed
to reused some of the debugging infrastructure) because these mappers
can't actually be switched.

Unfortunately this implementation detail did leak out. Till yesterday I
thought it leaked because we inherited from `Rom8kBBlocks` and that then
announces a 8kB block size. But I was mistaken, and MBilderbeek pointed
out this info is (independently from the class hierarchy) stored in
`romTypeInfoArray` in `RomInfo.cc`.

So all this patch does is correct the info in that table.