Prefer std::array over plain-C arrays
The former more easily integrates in c++ and can be passed by value
(does't decay to a pointer) etc.

In many cases the difference doesn't matter, but it's good to be
consistent.

We already followed this rule in most of the code. This patch also fixes
it in some windows specific code (which I can't easily test myself).
Though if I made a mistake it will manifest itself as a compile error
(not a bug at runtime) and then the github build action will inform me.