Switch from 'l' to 'L' literal suffix
And:
 ll  -> LL
 ul  -> UL
 ull -> ULL

My preference was always for the lower case version. But 'sonarcloud'
suggested to change it. The rationale is that the digit '1' and the
lower case 'l' look very similar (of course depending on the exact font
you're using). And then this might get confusing for literals like
'10l'. That argument convinced me. Hence this patch