Use transparent comparator in std::map
Replace
    std::map<K, V>      // 3rd template parameter defaults to std::less<K>
with
    std::map<K, V, std::less<>>

See here for why the latter is 'better':
    https://stackoverflow.com/questions/20317413/what-are-transparent-comparators

For example with:
    std::map<std::string, int> myMap;
    ...
    myMap.find("abc");
The find() method would internally construct a std::string object from
the string literal (const char*) "abc". If we instead define the map as:
    std::map<std::string, int, std::less<>> myMap;
then the find() method can directly compare std::string objects with
string-literals. In other words: it can perform heterogeneous
comparisons.

It's unfortunate that heterogeneous comparisons can't be done with the
shorter syntax. See the above stackoverflow post for the reasoning (in
short: backwards compatibility between c++11 and c++14).

--

This patch changes all std::map's in openMSX so that heterogeneous
comparisons can be used. Though at first sight I couldn't see any
opportunities where this could be exploited. But it's good to be
prepared anyway.