Replace our 'ranges::sort()' with 'std::ranges::sort()'
This is an experiment to check that all compilers we're using now
support std::ranges. I think it is after we've switched to clang-15. If
successful we can replace many more of our reimplementations.

'std::ranges::sort()' should be a drop-in replacement for our
'ranges::sort()'. We only need to add '#include <algorithm>' (and
possibly we can remove '#include "ranges.hh"').

It's common to abbreviate long c++ namespace names. For example
'std::filesystem' is almost universally abbreviated to just 'fs' (via
namespace fs = std::filesystem). For 'std::ranges' there is no single
dominant abbreviation. But 'rg', 'rng' and 'ranges' are common ones.
* Personally I think 'rng' can easily be confused with 'random number
  generator', so I don't like it.
* 'ranges' would be my preference. But in our current code we already
  have a 'ranges' namespace. The goal is to eliminate it (replace it
  with standard stuff). But before that's done we cannot reuse 'ranges'
  as an abbreviation.
* That leaves 'rg'. So let's pick that, at least for now.