Added function_ref as a more lightweight version of std::function
Modeled after std::function_ref which is proposed for c++26.
  https://en.cppreference.com/w/cpp/utility/functional/function_ref

In short: 'function_ref' is a non-owning, non-nullable version of
std::function. Because of these limitations 'function_ref' can be more
lightweight than 'std::function'. For example 'std::function' may
sometimes require dynamic memory allocation, but 'function_ref' never
needs to allocate memory.

Specifically on my platform (Linux, gcc, libstdc++):
* sizeof(std::function) == 48
* sizeof(function_ref) == 16
This makes passing 'function_ref' objects around much cheaper than
'std::function'. (In fact that was the original motivation for this
patch).

The cost of invoking a 'function_ref' vs a 'std::function' should be
about the same, I think.

The main limitation is that 'function_ref' is non-owning. In practice
this means that the function (or lambda, or function object) that
'function_ref' refers to must outlive the 'function_ref' object. So e.g.
when passing a 'function_ref' as a parameter, that parameter must not be
stored and used later after the function finishes.