More tweaks in Shortcut code
* Renamed 'Shortcuts::Data' to 'Shortcuts::Shortcut'
This better describes what the data is about. Just 'Data' is a
sufficiently descriptive name inside the 'Shortcuts' class itself. But
this struct is also exposed in the public API (and even more after this
patch).

* Several methods take a group of 3 parameters:
     keyChord, local/global, repeat
Replace that with a single 'Shortcut' parameter. For example this allows
to write:
    setShortcut(i, Shortcut{});
to clear a specific shortcut.

* Another example of the above was the recently added checkShortcut()
function. For example:
    if (shortcuts.checkShortcut(ImGuiKey_End, LOCAL, false)) { ...
I suspect that in the future we will add many more of these calls. And
in fact my main motivation for this patch was to try to simplify such
calls. After this patch it's now possible to write the common case (=
just a key, other properties default) like:
    if (shortcuts.checkShortcut({ImGuiKey_End})) { ...
And if we do need special properties we can use designated initializer
syntax to only specify those properties that are different from the
default:
    if (shortcuts.checkShortcut({.keyChord = ImGuiKey_Up_Arrow, .repeat = true})) { ...  // note no need to specify .type = LOCAL

* I've simplified the API of the Shortcuts class a bit:
** setShortcut() can no longer partially update a shortcut (=only set
some of the properties)
** I removed setDefaultShortcut(), it was only used from one location,
and it can now easily be replaced with:
    setShortcut(i, getDefaultShortcut(i))