Add control for auto_enable_reverse
Some people don't want to start up with reverse, thats's why we have that
setting. Let's make it avaiable in the GUI.

Also simplified the help text a bit, to make it also suitable for display on
the GUI. I guess this should have been done with the previous commit, where it
was changed into a boolean setting.