sonarcloud: use init-statement in 'if' and 'for'
This is partly an experiment. The resulting code looks sometimes a bit
weird, but that might just be because I'm not used to this new syntax
yet. I certainly do understand the technical benefits.

(Since c++17 already) it's allowed to initialize a variable inside the
condition of an if-statement.

Take for example this code:
    std::string name1 = getName(param1);
    if (!name1.empty()) {
        // do stuff with 'name1'
    }
    std::string name2 = getName(param2);
    if (!name2.empty()) {
        // do stuff with 'name2'
    }

With the new init stuff we can now also write it like:
    if (std::string name = getName(param1);
        !name.empty()) {
        // do stuff with 'name'
    }
    if (std::string name = getName(param2);
        !name.empty()) {
        // do stuff with 'name'
    }

The main difference is the scope of the old 'name1' and 'name2'
variables.
* Before they leaked out into the surrounding scope.
* Smaller scopes can improve readability: no need to scan the full
  outser scope to see if there are any other uses of the variable.
* Because of that they both needed a unique name, in the new version we
  just use 'name' two times.
* Because the scope is smaller, the lifetime is also shorter, so in case
  of std::string, the compiler can destroy the objects sooner (which
  might result in more efficient code).
* And because the lifetime doesn't overlap anymore, the compiler may
  even reuse the same memory (stack slot or even heap memory).

But as I said at the top: this is an experiment. It's easy to over-use
this feature. So this patch does certainly not apply this transformation
in all locations where sonarcloud says (correctlty) that it's
technically possible to use it.