sonarcloud: follow 'the-rule-of-5'
When one or more of the follow is defined, all 5 should be defined:
* copy-constructor
* move-constructor
* copy-assignment-operator
* move-assignment-operator
* destructor

Prior to c++11 there were no move-operations yet, and then this was
known as 'the-rule-of-3'.

For backwards compatibility with c++98 code, when you define a
copy-constructor or copy-assignment, the compiler will not automatically
generate the corresponding move-variants. So technically when you delete
the copy-constructor/assignment it's not required to also delete the
move-variant. Prior to this patch much of the openMSX code exploited
that loophole. Though it's much clearer to explicitly write the
move-variants as well.