Shortcuts: don't allow to configure 'repeat' on GUI shortcuts
I believe that for all GUI shortcuts we define in the code, we should be
able to (statically) decide whether repeating the shortcut makes sense
or not. For example "CTRL+G" (= enter a new address in the hex editor)
does not make sense to repeat. And 'F7' (= step in debugger) can be
useful to repeat (and obviously doing only a single step remains
possible).

If my assumption is correct then we could/should simplify the shortcut
configuration: the user should no longer have to configure the 'repeat'
property in the "edit shortcut" dialog. And we should also no longer
save/restore this property in "settings.xml".

If I'm mistaken, if we do sometimes want to have this configurable, then
it shouldn't be too hard to revert this commit.