Shortcuts: tweak edit shortcut dialog
Tweak how the edit shortcuts GUI dialogs look and behave (no changes to
the shortcut behavior themselves).

Remove the 'edit' button, instead made the 'current shortcut' text
clickable (turned it into a button).

Clicking the '<key>', 'Restore default' and 'Set none' buttons also
closes the "Edit shortcut" dialog.

With this change I'm happy with the shortcut-infrastructure. The next
step would be to actually start using it in the code. (And then maybe
we'll find out we need more tweaks).