ImGuiDebugger: setup some shortcuts
For:
* step-in          F7
* step-over        F8
* step-out         Shift+F7
* step-back        Shift+F8
* break/continue   F5

By default these are local to both the disassembly and the
debugger-toolbar window. But some users may want to reconfigure them as
global, so active whenever ImGui has focus. I don't think these should
be global by default because then the F5/F7/F8 keys can no longer be
used for other (local) shortcuts (and not that many openMSX users will
(often) use the debugger).

In theory it's possible to create hotkeys for these actions, so that
they even work when the MSX window has focus. But personally I don'
think that will be needed: that's was also not possible with the old
(external) debugger (and nobody complained about that).

This is just my first idea for a possible key-assignment. Absolutely
feel free to tweak these. I'd also appreciate help with defining more
shortcuts for various other GUI actions.