AY8910: replace global variable with a global constant
For the 'detune' AY8910 feature we need an array of 256 random floats.
Currently that's a global array that's lazily initialized. In this patch
I'm experimenting with initializing that array at compile-time (see
links in the code for more info). For this usecase it's fine that the
random values are the same in each openMSX run. And this makes this
global _variable_ a global _constant_.