TC8566AF: read/write sector, detect missing sector only after 2 index pulses
Before this patch, when trying to read/write a non-existing sector, the
TC8566AF emulation would _immediately_ report an error. Though in
reality it takes some time before the FDC knows the sector is missing.
More specifically the error gets detected when the index pulse is seen
twice (so that we've certainly seen one full disk revolution). This
patch adds that proper timing.

Reading a non-existing sector is a nice trick to locate the index pulse
on the disk. Because, in contrast to the WD2793, the TC8566AF does not
allow to poll the index pulse via some status register.

Marcos Daniel Blanco Oliveira requested this improvement because he'd
like to use this trick in DSKPRO.